don't crash if mojang mapped jar doesn't exist yet

+++ subprojects/server.gradle.kts
@@ -16,16 +16,21 @@ dependencies {
     loadDependencies(File(project.projectDir, "pom.xml"), project, true)
 
     val projectDir = project.parent!!.projectDir
+    val og = project.parent!!.projectDir.resolve("work/1.15.2-mojang-mapped.jar")
+    if (!og.exists()) {
+        logger.warn("work/1.15.2-mojang-mapped.jar doesn't exist (yet)")
+        return@dependencies
+    }
     for (i in 0..100) {
         val file = projectDir.resolve("work/1.15.2-mojang-mapped-copied-$i.jar")
         if (file.exists()) {
             if (file.delete()) {
-                project.parent!!.projectDir.resolve("work/1.15.2-mojang-mapped.jar").copyTo(file, true)
+                og.copyTo(file, true)
                 api(files(file))
                 break
             }
         } else {
-            project.parent!!.projectDir.resolve("work/1.15.2-mojang-mapped.jar").copyTo(file, true)
+            og.copyTo(file, true)
             api(files(file))
             break
         }
